+++ src/format.task.ts
@@ -3,8 +3,8 @@ import {printSpawnResult} from '@feltcoop/felt/util/process.js';
 import {TaskError, type Task} from './task/task.js';
 import {formatDirectory} from './format/formatDirectory.js';
 import {paths} from './paths.js';
-import {type FormatTaskArgs} from './format.js';
-import {FormatTaskArgsSchema} from './format.schema.js';
+import {type FormatTaskArgs} from './formatTask.js';
+import {FormatTaskArgsSchema} from './formatTask.schema.js';
 
 export interface TaskArgs {
 	check?: boolean;
