+++ src/gen.task.ts
@@ -12,8 +12,8 @@ import {formatFile} from './format/formatFile.js';
 import {printPath} from './paths.js';
 import {loadConfig} from './config/config.js';
 import {buildSource} from './build/buildSource.js';
-import {type GenTaskArgs} from './gen.js';
-import {GenTaskArgsSchema} from './gen.schema.js';
+import {type GenTaskArgs} from './genTask.js';
+import {GenTaskArgsSchema} from './genTask.schema.js';
 
 // TODO test - especially making sure nothing gets genned
 // if there's any validation or import errors
