+++ src/lintTask.ts
